Add interactive list in state module
