+++ src/index.js
@@ -1,5 +1,6 @@
 import './style.css';
 import display from './ui.js';
 import './user.js';
+import './state.js'
 
 document.addEventListener('DOMContentLoaded', display);
